requirements.txt
openai
requests
semgrep
python-dotenv
jsonschema
requests